refactor: move profile api and sales-summary api into myself handler
