+++ src/main/kotlin/com/kroffle/knitting/controller/filter/auth/AuthorizationFilter.kt
@@ -11,9 +11,9 @@ import org.springframework.web.server.WebFilterChain
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import com.kroffle.knitting.controller.router.auth.LogInRouter.Companion.PUBLIC_PATHS as LogInRouterPublicPaths
-import com.kroffle.knitting.controller.router.auth.ProfileRouter.Companion.PUBLIC_PATHS as ProfileRouterPublicPaths
 import com.kroffle.knitting.controller.router.design.DesignRouter.Companion.PUBLIC_PATHS as DesignRouterPublicPaths
 import com.kroffle.knitting.controller.router.design.DesignsRouter.Companion.PUBLIC_PATHS as DesignsRouterPublicPaths
+import com.kroffle.knitting.controller.router.knitter.MyselfRouter.Companion.PUBLIC_PATHS as MyselfRouterPublicPaths
 import com.kroffle.knitting.controller.router.ping.PingRouter.Companion.PUBLIC_PATHS as PingRouterPublicPaths
 import com.kroffle.knitting.controller.router.product.ProductRouter.Companion.PUBLIC_PATHS as ProductRouterPublicPaths
 
@@ -66,7 +66,7 @@ class AuthorizationFilter(private val tokenDecoder: TokenDecoder) : WebFilter {
         private const val HEADER_PREFIX = "Bearer "
         private val PUBLIC_PATHS = (
             LogInRouterPublicPaths +
-                ProfileRouterPublicPaths +
+                MyselfRouterPublicPaths +
                 DesignRouterPublicPaths +
                 DesignsRouterPublicPaths +
                 PingRouterPublicPaths +
