+++ src/main/kotlin/com/kroffle/knitting/controller/handler/knitter/dto/MyProfileResponse.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.controller.handler.auth.dto
+package com.kroffle.knitting.controller.handler.knitter.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.kroffle.knitting.controller.handler.helper.response.type.ObjectPayload
